v1.3.0
It was more than 2 months gap since the last release and here we are with
a couple of bug fixes and better ES2015 support.

A big part of fixes related to `requireDescriptionCompleteSentence` rule.
If you had troubles with it you can give it a try now.

`enforceExistence` now supports ES2015 exports and treats it like usual
CommonJS `module.exports`.

Also we should note that [Jordan Harband](http://github.com/ljharb) came
with a `node v0.8` support :tada:. Thank you, Jordan!

See all changes in CHANGELOG.md

Thank you for usign it and contributing to it!